Change vulkan particle fields from double to float
I get an error from a layer about doubles, so use floats instead to make
sure all gpus can run this

+++ vulkanSetup.c
@@ -30,7 +30,7 @@ typedef struct vertex_t {
 	float z;
 } vertex;
 typedef struct {
-	double posX, posY, dirX, dirY, angle;
+	float posX, posY, dirX, dirY, angle;
 } particle;
 extern const int particleCount;
 VkBuffer vertexBuf, particleBuf;
